+++ packages/runtime/tce-edit-vue3/src/main.ts
@@ -7,11 +7,9 @@ import Radio from './radio';
 import vuetify from './plugins/vuetify';
 
 const element = await import(import.meta.env.EDIT_DIR);
-const isQuestion = !!element.default.isQuestion;
-const isComposite = !!element.default.isComposite;
-const gradingType = element.default.gradingType;
+const { isComposite = false, isQuestion, isGradable } = element.default;
 
-const app = createApp(App, { isQuestion, gradingType });
+const app = createApp(App, { isQuestion, isGradable });
 const radio = Radio.getInstance();
 app.provide('$eventBus', radio);
 app.provide('$elementBus', radio.channel('app'));